feat: Add file selector plugins for Linux and Windows
This commit adds the file selector plugins for Linux and Windows platforms. The Linux plugin is added in the `generated_plugins.cmake` file, and the Windows plugin is added in both the `generated_plugins.cmake` and `generated_plugin_registrant.cc` files. This allows the application to select files from the file system on both platforms.

The following files are modified:
- `linux/flutter/generated_plugins.cmake`
- `windows/flutter/generated_plugins.cmake`
- `linux/flutter/generated_plugin_registrant.cc`
- `windows/flutter/generated_plugin_registrant.cc`